Make it as individual class to make the codefactor happy.
